src/python/grapl-http-service/requirements.txt
pytest
pytest-cov
hypothesis
pyjwt
chalice
boto3_stubs
